fix(api): generalize workflow template orchestration
