+++ apps/api/tests/test_workflow_registry.py
@@ -1,0 +1,150 @@
+from __future__ import annotations
+
+from typing import cast
+
+import pytest
+
+from noa_api.core.tool_error_sanitizer import SanitizedToolError
+from noa_api.core.workflows.registry import (
+    build_approval_context,
+    build_workflow_todos,
+    describe_workflow_activity,
+    fetch_postflight_result,
+    infer_waiting_on_user_workflow_from_messages,
+    list_registered_workflow_families,
+    register_workflow_template,
+    require_matching_preflight,
+)
+from noa_api.core.workflows.types import (
+    WorkflowInference,
+    WorkflowTemplate,
+    WorkflowTemplateContext,
+)
+
+
+class _CustomWorkflowTemplate(WorkflowTemplate):
+    def build_todos(self, context: WorkflowTemplateContext):
+        return [
+            {
+                "content": f"Registered step for {context.tool_name}.",
+                "status": "waiting_on_approval",
+                "priority": "high",
+            }
+        ]
+
+    def describe_activity(
+        self, *, tool_name: str, args: dict[str, object]
+    ) -> str | None:
+        _ = tool_name, args
+        return "Run custom workflow"
+
+    def build_before_state(
+        self,
+        *,
+        tool_name: str,
+        args: dict[str, object],
+        preflight_results: list[dict[str, object]],
+    ) -> list[dict[str, str]] | None:
+        _ = tool_name, args, preflight_results
+        return [{"label": "Before", "value": "ready"}]
+
+    def require_preflight(
+        self,
+        *,
+        tool_name: str,
+        args: dict[str, object],
+        working_messages: list[dict[str, object]],
+        requested_server_id: str | None,
+    ) -> SanitizedToolError | None:
+        _ = tool_name, args, working_messages, requested_server_id
+        return SanitizedToolError(
+            error="Custom preflight missing",
+            error_code="custom_preflight_required",
+        )
+
+    async def fetch_postflight_result(
+        self, *, tool_name: str, args: dict[str, object], session
+    ):
+        _ = tool_name, args, session
+        return {"ok": True, "source": "custom-template"}
+
+    def infer_waiting_on_user_workflow(
+        self,
+        *,
+        assistant_text: str,
+        working_messages: list[dict[str, object]],
+    ) -> WorkflowInference | None:
+        _ = working_messages
+        if "reason" not in assistant_text.lower():
+            return None
+        return WorkflowInference(tool_name="custom_change", args={"resource": "alpha"})
+
+
+@pytest.mark.asyncio
+async def test_workflow_registry_supports_registered_templates() -> None:
+    family = "test-custom-workflow-family"
+    register_workflow_template(family=family, template=_CustomWorkflowTemplate())
+
+    assert family in list_registered_workflow_families()
+
+    todos = build_workflow_todos(
+        tool_name="custom_change",
+        workflow_family=family,
+        args={"resource": "alpha"},
+        phase="waiting_on_approval",
+        preflight_evidence=[],
+    )
+    assert todos == [
+        {
+            "content": "Registered step for custom_change.",
+            "status": "waiting_on_approval",
+            "priority": "high",
+        }
+    ]
+
+    approval_context = build_approval_context(
+        tool_name="custom_change",
+        workflow_family=family,
+        args={"resource": "alpha"},
+        working_messages=[],
+    )
+    assert approval_context["activity"] == "Run custom workflow"
+    assert approval_context["beforeState"] == [{"label": "Before", "value": "ready"}]
+    assert approval_context["argumentSummary"] == [
+        {"label": "Resource", "value": "alpha"}
+    ]
+
+    preflight_error = require_matching_preflight(
+        tool_name="custom_change",
+        workflow_family=family,
+        args={"resource": "alpha"},
+        working_messages=[],
+    )
+    assert preflight_error is not None
+    assert preflight_error.error_code == "custom_preflight_required"
+
+    postflight_result = await fetch_postflight_result(
+        tool_name="custom_change",
+        workflow_family=family,
+        args={"resource": "alpha"},
+        session=cast(object, object()),
+    )
+    assert postflight_result == {"ok": True, "source": "custom-template"}
+
+    inferred = infer_waiting_on_user_workflow_from_messages(
+        assistant_text="I still need a reason before I can continue.",
+        working_messages=[],
+    )
+    assert inferred == WorkflowInference(
+        tool_name="custom_change",
+        args={"resource": "alpha"},
+    )
+
+    assert (
+        describe_workflow_activity(
+            tool_name="custom_change",
+            workflow_family=family,
+            args={"resource": "alpha"},
+        )
+        == "Run custom workflow"
+    )
